requirements.txt
beautifulsoup4==4.11.2
lxml==4.9.2
PyMySQL==1.0.2
python-dotenv==1.0.0
soupsieve==2.4
